chore: enhance justfile with testing commands and install git hooks
This improves the development workflow by facilitating automatic code formatting and avoiding the time-consuming process of waiting for the CI format check job to fail, then running Rustfmt and pushing the formatted code.

* Added commands for running unit and integration tests in the `justfile`.
* Introduced a script to install pre-commit hooks, ensuring code quality checks.
* Created a new `.github/pre-commit-config.yaml` file to define formatting hooks for Rust code.

Note that git hooks are a convenient way of running rustfmt before committing anything. But they are opt-in. Run the `just install-git-hooks` task to install the [pre-commit](https://pre-commit.com/) hooks

Signed-off-by: Lorenzo Delgado <lorenzo@edgeandnode.com>