refactor: rename ScannedRange to ParquetMeta (#279)
This resolves some naming and documentation issues associated with the lingering "scanned ranges". This previously referred to multiple things including the metadata stored in the parquet file, the block range associated with the file, or an entry in a meta table which has been replaced by the metadata DB. The name `ParquetMeta` was chosen due to it's similarity to `ObjectMeta`.