add BlockRange and fork choice algorithm (#343)
This introduces the `BlockRange` struct required in the metadata for a parquet file (in the file and in the metadata DB). This also introduces a `TableRanges` struct, which allows a table writer to incrementally update which parquet files are associated with the canonical chain for future readers. The `TableRanges::update` algorithm supports out-of-order insertion of block ranges, since we expect ranges to come in concurrently.