test reorg handling against anvil RPC (#398)
This adds a test that spins up an anvil node, which gives us control over reorgs without having to push testing concerns into `dump` to get control over the `BlockStreamer`. In the future this will cover streaming SQL responses from the JSONL server as reorgs occur. For now, this test just makes sure the metadata output of dump is written as expected when a reorg occurs.