fix(dump): split segment on reorg (#499)
This maintains a precondition for segments that all blocks within a `BlockRange` for some segment form a valid chain. We can maintain this invariant in the `RawTableWriter` if we split the segment when we detect a hash mismatch between the prior block range and the incoming block range. We have already been splitting segments when the partition size is exceeded.