rewrite segments (#526)
This rewrites the contents of `segments.rs` to make the implementation deterministic even when compaction happens. This new implementation uses the object's timestamp to prioritize most recent segments when searching for the canonical & fork chains. This avoids a lot of implementation issues, and works assuming that compaction crates objects with more recent timestamps. This also feels more intuitive for the user that new segments override old ones, instead of a set of rules that are difficult to reason about.