combine streaming watermarks and execution loop (#554)
This fixes #549 by removing the intermediate streams between source table watermarks and the streaming execution loop. This allows us to resolve arbitrary block ranges, even when the executable range is truncated to the microbatch limit, and pass them to the consumer.

The next PR will resolve the TODO for handling reorgs before the previous range.