make SQL table dump linear (#726)
This changes SQL dumps to be linear. The single resuming streaming query is resumed using the `BlockRange` of the most recently dumped segment of the destination `PhysicalTable`. This limits the use of `PhysicalTable::missing_ranges` to dumping raw datasets. As an added bonus, reorgs on SQL datasets are now resolved in a single dump (rather than one additional dump per block of reorg depth).

Resolves #567.