fix(extractors): remove EVM-RPC batched block stream retry logic (#865)
- Recently a wrapper around block streams was added that introduces
generic retry logic when a stream returns an error. Therefore, the retry
logic specific to the batched EVM-RPC block stream is redundant.
- Additionaly, this retry logic had a bug where upon batch failure,
JSON-RPC calls were retried individually and (on success) the results
appended to aggregate collection but then the entire batch (even the
calls that succeeded individually) was inserted back into the collection
of all calls, which lead to writing the same block twice in certain
cases.