fix: send schema message on first microbatch start (#893)
This fixes a bug where, if the first microbatch yields no batches then the metadata message on microbatch end would be sent out before the schema. This breaks arrow flight clients which expect the schema to come first. Instead, the schema message is always sent on the first microbatch start.