fix(datasets-derived): allow fully-qualified dataset references in SQL (#1207)
This is a temporary measure to allow derived dataset manifests with fully qualified dataset references in the SQL queries. But I think the correct user experience would be to resolve the unqualified name into the qualified one mapped in the dependencies section of the manifest. Alternatively, we should make the dependencies section an array to remove the implication that the dependency could be aliased.